refactor: changes rating type to BigDecimal, refactors GET files/id endpoint to return raw image and content-type

+++ main/src/main/java/ru/practicum/ewm/dto/event/EventFullDto.java
@@ -7,6 +7,8 @@ import ru.practicum.ewm.dto.category.CategoryDto;
 import ru.practicum.ewm.dto.location.LocationDto;
 import ru.practicum.ewm.dto.user.UserShortDto;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -95,5 +97,5 @@ public class EventFullDto {
     /*
      * Рейтинг события
      */
-    private double rating;
+    private BigDecimal rating;
 }
