+++ main/src/main/java/ru/practicum/ewm/repository/event/EventRepository.java
@@ -12,6 +12,7 @@ import ru.practicum.ewm.model.event.Event;
 import ru.practicum.ewm.model.event.EventState;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,7 +34,7 @@ public interface EventRepository extends JpaRepository<Event, Long> {
     @Modifying
     @Transactional
     @Query(value = "UPDATE events SET rating = :rating WHERE event_id = :eventId", nativeQuery = true)
-    void updateEventRating(Double rating, Long eventId);
+    void updateEventRating(BigDecimal rating, Long eventId);
 
     default Event getExistingEvent(Long eventId) {
         return findById(eventId).orElseThrow(() -> {
